requirements.txt
flowserv-core>=0.3.2
streamlit
